+++ CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to this library are documented in this file.
 
 ### Bug Fixes
 
+- Add preflight check of MOS database write for known column storage.
 - Support `with_sqlalchemy_conn` to decorate a generator.
 
 ## **1.27.0** (14 Apr 2026)